refactor(ai): make ChatMessageAdaptor a static util in internals package
